Move FileHandler to filehandler.py
